refactor dpm portfolio memory search route
